Updated calabash version and project settings to get acceptance tests to run again. NOTE: also required PC restart and possibly some other unknown magic

+++ acceptanceTests.sh
@@ -3,29 +3,53 @@
 FEATURE_FILES=${1}
 
 CURRENT_DIR=`pwd`
-BUILD_DIR="$CURRENT_DIR/build_acceptanceTest"
+BUILD_DIR="$CURRENT_DIR/build"
+BUILD_OUTPUT_DIR="$BUILD_DIR/Build/Products/Debug-iphonesimulator"
 APP_NAME="TheWriteTrack"
+SIM_DEVICE='iPhone 6'
+OS_VERSION='9.0'
+PLATFORM='iOS Simulator'
 
-killall 'iOS Simulator' 1&>/dev/null
 
-rm -rf $BUILD_DIR
+function stop_sim_if_necessary {
+    if [[ "$PLATFORM" == "iOS Simulator" ]]; then
+        killall 'iOS Simulator' 1&>/dev/null
+    fi
+}
 
-bundle install
+function clean_environment {
+    rm -rf $BUILD_DIR
+    bundle install
+    bundle exec pod install
+    bundle exec calabash-ios sim reset
+}
 
-bundle exec pod install
-
-xctool \
+function run_tests {
+    xcrun xcodebuild \
+     -derivedDataPath "${BUILD_DIR}" \
+     ARCHS="i386 x86_64" \
+     VALID_ARCHS="i386 x86_64" \
+     ONLY_ACTIVE_ARCH=NO \
      -workspace $APP_NAME.xcworkspace/ \
-	 -scheme $APP_NAME-cal \
-     CONFIGURATION_BUILD_DIR=$BUILD_DIR \
-	 -destination "platform=iOS Simulator,name=iPhone 6,OS=8.3" \
-     -configuration Debug \
+     -scheme $APP_NAME-cal \
      -sdk iphonesimulator \
-	 clean \
-	 build
+     -configuration Debug \
+     clean build
 
-bundle exec calabash-ios sim reset
+    export DEVICE_TARGET=`instruments -s devices | grep -v "+" | grep "${SIM_DEVICE} (${OS_VERSION})"`
+    export APP="$BUILD_DIR/Build/Products/Debug-iphonesimulator/$APP_NAME.app"
+    bundle exec "DEBUG=1 cucumber ${FEATURE_FILES}"
+}
 
-bundle exec "APP_BUNDLE_PATH=$BUILD_DIR/$APP_NAME.app TARGET_DEVICE='iPhone 6 (iOS 8.3)' cucumber ${FEATURE_FILES}"
+#______EXECUTE_________________________
 
-killall 'iOS Simulator' 1&>/dev/null
+stop_sim_if_necessary
+
+clean_environment
+
+run_tests
+
+stop_sim_if_necessary
+
+#______FIN______________________________
+
